fix(web): keep db-only modules out of client bundles
