+++ apps/web/src/app/agents/team/AgentPnL.tsx
@@ -13,7 +13,7 @@ import { Switch } from '@/components/ui/switch';
 import { useAuth } from '@/hooks/useAuth';
 import { useCollapsibleHeight } from '@/hooks/useCollapsibleHeight';
 import { usePortfolioPnL } from '@/hooks/usePortfolioPnL';
-import type { UserPositionsSnapshot } from '@/lib/markets/user-positions';
+import type { UserPositionsSnapshot } from '@/lib/markets/user-positions-types';
 import {
   usePerpPositions,
   usePredictionPositions,
