+++ apps/web/src/components/wallet/v2/pnl-tab.tsx
@@ -5,7 +5,7 @@ import { ChevronDown } from 'lucide-react';
 import { useCallback, useMemo, useRef, useState } from 'react';
 import { useOnClickOutside } from '@/hooks/useOnClickOutside';
 import type { TeamTradingSummary } from '@/lib/agents/team-trading-summary';
-import type { PnlHistoryScope } from '@/lib/wallet/pnlHistory';
+import type { PnlHistoryScope } from '@/lib/wallet/pnl-history-types';
 import { PnLChart } from './pnl-chart';
 
 interface PnLTabProps {
