+++ apps/web/src/hooks/useAgentsTeamDashboard.ts
@@ -2,7 +2,7 @@
 
 import { logger } from '@babylon/shared';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import type { TeamDashboardAgent } from '@/lib/agents/team-dashboard';
+import type { TeamDashboardAgent } from '@/lib/agents/team-dashboard-types';
 import type { TeamTradingSummary } from '@/lib/agents/team-trading-summary';
 
 interface TeamDashboardResponse {
